Reorganize schema class hierarchy with respect to schema object names
Currently, the base `Object` class assumes that the object name is
fully-qualified, and then there are a bunch of exceptions down the
inheritance hierarchy that override this assumption.  This inversion
arrangement is unsound, so reorganize the hierarchy such that the
name in `schema.Object` is considered unqualified, and explicit
inheritance from `schema.QualifiedObject` is needed to enable
module-qualified behavior for object names.