Adjust the DESCRIBE AS SDL and AS TEXT logic.
Change the logic when DESCRIBE command shows the original expression.
Basically if the normalized expression appears to only introduce a
default module that is the same as the module the object belongs to then
SDL and TEXT versions of describe can safely use the original short
expression. Otherwise, a normalized expression must be given to avoid
confusion.

Fixes: #1254